refactor: move Tasklist related templates to template package
