+++ search/search-client-query-transformer/src/main/java/io/camunda/search/clients/transformers/filter/UserTaskFilterTransformer.java
@@ -16,23 +16,23 @@ import static io.camunda.search.clients.query.SearchQueryBuilders.intOperations;
 import static io.camunda.search.clients.query.SearchQueryBuilders.longTerms;
 import static io.camunda.search.clients.query.SearchQueryBuilders.stringOperations;
 import static io.camunda.search.clients.query.SearchQueryBuilders.stringTerms;
-import static io.camunda.webapps.schema.descriptors.tasklist.template.TaskTemplate.ASSIGNEE;
-import static io.camunda.webapps.schema.descriptors.tasklist.template.TaskTemplate.BPMN_PROCESS_ID;
-import static io.camunda.webapps.schema.descriptors.tasklist.template.TaskTemplate.CANDIDATE_GROUPS;
-import static io.camunda.webapps.schema.descriptors.tasklist.template.TaskTemplate.CANDIDATE_USERS;
-import static io.camunda.webapps.schema.descriptors.tasklist.template.TaskTemplate.COMPLETION_TIME;
-import static io.camunda.webapps.schema.descriptors.tasklist.template.TaskTemplate.CREATION_TIME;
-import static io.camunda.webapps.schema.descriptors.tasklist.template.TaskTemplate.DUE_DATE;
-import static io.camunda.webapps.schema.descriptors.tasklist.template.TaskTemplate.FLOW_NODE_BPMN_ID;
-import static io.camunda.webapps.schema.descriptors.tasklist.template.TaskTemplate.FLOW_NODE_INSTANCE_ID;
-import static io.camunda.webapps.schema.descriptors.tasklist.template.TaskTemplate.FOLLOW_UP_DATE;
-import static io.camunda.webapps.schema.descriptors.tasklist.template.TaskTemplate.IMPLEMENTATION;
-import static io.camunda.webapps.schema.descriptors.tasklist.template.TaskTemplate.KEY;
-import static io.camunda.webapps.schema.descriptors.tasklist.template.TaskTemplate.PRIORITY;
-import static io.camunda.webapps.schema.descriptors.tasklist.template.TaskTemplate.PROCESS_DEFINITION_ID;
-import static io.camunda.webapps.schema.descriptors.tasklist.template.TaskTemplate.PROCESS_INSTANCE_ID;
-import static io.camunda.webapps.schema.descriptors.tasklist.template.TaskTemplate.STATE;
-import static io.camunda.webapps.schema.descriptors.tasklist.template.TaskTemplate.TENANT_ID;
+import static io.camunda.webapps.schema.descriptors.template.TaskTemplate.ASSIGNEE;
+import static io.camunda.webapps.schema.descriptors.template.TaskTemplate.BPMN_PROCESS_ID;
+import static io.camunda.webapps.schema.descriptors.template.TaskTemplate.CANDIDATE_GROUPS;
+import static io.camunda.webapps.schema.descriptors.template.TaskTemplate.CANDIDATE_USERS;
+import static io.camunda.webapps.schema.descriptors.template.TaskTemplate.COMPLETION_TIME;
+import static io.camunda.webapps.schema.descriptors.template.TaskTemplate.CREATION_TIME;
+import static io.camunda.webapps.schema.descriptors.template.TaskTemplate.DUE_DATE;
+import static io.camunda.webapps.schema.descriptors.template.TaskTemplate.FLOW_NODE_BPMN_ID;
+import static io.camunda.webapps.schema.descriptors.template.TaskTemplate.FLOW_NODE_INSTANCE_ID;
+import static io.camunda.webapps.schema.descriptors.template.TaskTemplate.FOLLOW_UP_DATE;
+import static io.camunda.webapps.schema.descriptors.template.TaskTemplate.IMPLEMENTATION;
+import static io.camunda.webapps.schema.descriptors.template.TaskTemplate.KEY;
+import static io.camunda.webapps.schema.descriptors.template.TaskTemplate.PRIORITY;
+import static io.camunda.webapps.schema.descriptors.template.TaskTemplate.PROCESS_DEFINITION_ID;
+import static io.camunda.webapps.schema.descriptors.template.TaskTemplate.PROCESS_INSTANCE_ID;
+import static io.camunda.webapps.schema.descriptors.template.TaskTemplate.STATE;
+import static io.camunda.webapps.schema.descriptors.template.TaskTemplate.TENANT_ID;
 import static java.util.Optional.ofNullable;
 
 import io.camunda.search.clients.query.SearchQuery;
