+++ tasklist/els-schema/src/main/java/io/camunda/tasklist/store/elasticsearch/DraftVariablesStoreElasticSearch.java
@@ -17,7 +17,7 @@ import io.camunda.tasklist.exceptions.TasklistRuntimeException;
 import io.camunda.tasklist.store.DraftVariableStore;
 import io.camunda.tasklist.tenant.TenantAwareElasticsearchClient;
 import io.camunda.tasklist.util.ElasticsearchUtil;
-import io.camunda.webapps.schema.descriptors.tasklist.template.DraftTaskVariableTemplate;
+import io.camunda.webapps.schema.descriptors.template.DraftTaskVariableTemplate;
 import io.camunda.webapps.schema.entities.usertask.DraftTaskVariableEntity;
 import java.io.IOException;
 import java.util.Collection;
@@ -67,49 +67,22 @@ public class DraftVariablesStoreElasticSearch implements DraftVariableStore {
   @Qualifier("tasklistObjectMapper")
   private ObjectMapper objectMapper;
 
-  public void createOrUpdate(Collection<DraftTaskVariableEntity> draftVariables) {
+  @Override
+  public void createOrUpdate(final Collection<DraftTaskVariableEntity> draftVariables) {
     final BulkRequest bulkRequest = new BulkRequest();
-    for (DraftTaskVariableEntity variableEntity : draftVariables) {
+    for (final DraftTaskVariableEntity variableEntity : draftVariables) {
       bulkRequest.add(createUpsertRequest(variableEntity));
     }
     try {
       ElasticsearchUtil.processBulkRequest(
           esClient, bulkRequest, WriteRequest.RefreshPolicy.WAIT_UNTIL);
-    } catch (PersistenceException ex) {
+    } catch (final PersistenceException ex) {
       throw new TasklistRuntimeException(ex);
     }
   }
 
-  private UpdateRequest createUpsertRequest(DraftTaskVariableEntity draftVariableEntity) {
-    try {
-      final Map<String, Object> updateFields = new HashMap<>();
-      updateFields.put(DraftTaskVariableTemplate.TASK_ID, draftVariableEntity.getTaskId());
-      updateFields.put(DraftTaskVariableTemplate.NAME, draftVariableEntity.getName());
-      updateFields.put(DraftTaskVariableTemplate.VALUE, draftVariableEntity.getValue());
-      updateFields.put(DraftTaskVariableTemplate.FULL_VALUE, draftVariableEntity.getFullValue());
-      updateFields.put(DraftTaskVariableTemplate.IS_PREVIEW, draftVariableEntity.getIsPreview());
-
-      // format date fields properly
-      final Map<String, Object> jsonMap =
-          objectMapper.readValue(objectMapper.writeValueAsString(updateFields), HashMap.class);
-
-      return new UpdateRequest()
-          .index(draftTaskVariableTemplate.getFullQualifiedName())
-          .id(draftVariableEntity.getId())
-          .upsert(objectMapper.writeValueAsString(draftVariableEntity), XContentType.JSON)
-          .doc(jsonMap)
-          .retryOnConflict(UPDATE_RETRY_COUNT);
-
-    } catch (IOException e) {
-      throw new TasklistRuntimeException(
-          String.format(
-              "Error preparing the query to upsert task variable instance [%s]",
-              draftVariableEntity.getId()),
-          e);
-    }
-  }
-
-  public long deleteAllByTaskId(String taskId) {
+  @Override
+  public long deleteAllByTaskId(final String taskId) {
     final DeleteByQueryRequest request =
         new DeleteByQueryRequest(draftTaskVariableTemplate.getFullQualifiedName());
     request.setQuery(QueryBuilders.termQuery(DraftTaskVariableTemplate.TASK_ID, taskId));
@@ -117,7 +90,7 @@ public class DraftVariablesStoreElasticSearch implements DraftVariableStore {
     try {
       final BulkByScrollResponse response = esClient.deleteByQuery(request, RequestOptions.DEFAULT);
       return response.getDeleted(); // Return the count of deleted documents
-    } catch (IOException e) {
+    } catch (final IOException e) {
       throw new TasklistRuntimeException(
           String.format(
               "Error preparing the query to delete draft task variable instances for task [%s]",
@@ -126,8 +99,9 @@ public class DraftVariablesStoreElasticSearch implements DraftVariableStore {
     }
   }
 
+  @Override
   public List<DraftTaskVariableEntity> getVariablesByTaskIdAndVariableNames(
-      String taskId, List<String> variableNames) {
+      final String taskId, final List<String> variableNames) {
     try {
       final BoolQueryBuilder queryBuilder =
           QueryBuilders.boolQuery()
@@ -146,7 +120,7 @@ public class DraftVariablesStoreElasticSearch implements DraftVariableStore {
 
       return ElasticsearchUtil.scroll(
           searchRequest, DraftTaskVariableEntity.class, objectMapper, esClient);
-    } catch (IOException e) {
+    } catch (final IOException e) {
       throw new TasklistRuntimeException(
           String.format(
               "Error executing the query to get draft task variable instances for task [%s] with variable names %s",
@@ -155,7 +129,8 @@ public class DraftVariablesStoreElasticSearch implements DraftVariableStore {
     }
   }
 
-  public Optional<DraftTaskVariableEntity> getById(String variableId) {
+  @Override
+  public Optional<DraftTaskVariableEntity> getById(final String variableId) {
     try {
       final SearchRequest searchRequest =
           new SearchRequest(draftTaskVariableTemplate.getFullQualifiedName());
@@ -175,7 +150,7 @@ public class DraftVariablesStoreElasticSearch implements DraftVariableStore {
       final DraftTaskVariableEntity entity =
           objectMapper.readValue(sourceAsString, DraftTaskVariableEntity.class);
       return Optional.of(entity);
-    } catch (IOException e) {
+    } catch (final IOException e) {
       LOGGER.error(
           String.format("Error retrieving draft task variable instance with ID [%s]", variableId),
           e);
@@ -184,7 +159,7 @@ public class DraftVariablesStoreElasticSearch implements DraftVariableStore {
   }
 
   @Override
-  public List<String> getDraftVariablesIdsByTaskIds(List<String> taskIds) {
+  public List<String> getDraftVariablesIdsByTaskIds(final List<String> taskIds) {
     final SearchRequest searchRequest =
         new SearchRequest(draftTaskVariableTemplate.getFullQualifiedName())
             .source(
@@ -193,8 +168,37 @@ public class DraftVariablesStoreElasticSearch implements DraftVariableStore {
                     .fetchField(DraftTaskVariableTemplate.ID));
     try {
       return ElasticsearchUtil.scrollIdsToList(searchRequest, esClient);
-    } catch (IOException e) {
+    } catch (final IOException e) {
       throw new TasklistRuntimeException(e.getMessage(), e);
+    }
+  }
+
+  private UpdateRequest createUpsertRequest(final DraftTaskVariableEntity draftVariableEntity) {
+    try {
+      final Map<String, Object> updateFields = new HashMap<>();
+      updateFields.put(DraftTaskVariableTemplate.TASK_ID, draftVariableEntity.getTaskId());
+      updateFields.put(DraftTaskVariableTemplate.NAME, draftVariableEntity.getName());
+      updateFields.put(DraftTaskVariableTemplate.VALUE, draftVariableEntity.getValue());
+      updateFields.put(DraftTaskVariableTemplate.FULL_VALUE, draftVariableEntity.getFullValue());
+      updateFields.put(DraftTaskVariableTemplate.IS_PREVIEW, draftVariableEntity.getIsPreview());
+
+      // format date fields properly
+      final Map<String, Object> jsonMap =
+          objectMapper.readValue(objectMapper.writeValueAsString(updateFields), HashMap.class);
+
+      return new UpdateRequest()
+          .index(draftTaskVariableTemplate.getFullQualifiedName())
+          .id(draftVariableEntity.getId())
+          .upsert(objectMapper.writeValueAsString(draftVariableEntity), XContentType.JSON)
+          .doc(jsonMap)
+          .retryOnConflict(UPDATE_RETRY_COUNT);
+
+    } catch (final IOException e) {
+      throw new TasklistRuntimeException(
+          String.format(
+              "Error preparing the query to upsert task variable instance [%s]",
+              draftVariableEntity.getId()),
+          e);
     }
   }
 }
