+++ tasklist/els-schema/src/test/java/io/camunda/tasklist/store/opensearch/FormStoreOpenSearchTest.java
@@ -20,7 +20,7 @@ import io.camunda.tasklist.store.FormStore.FormIdView;
 import io.camunda.tasklist.tenant.TenantAwareOpenSearchClient;
 import io.camunda.webapps.schema.descriptors.index.FormIndex;
 import io.camunda.webapps.schema.descriptors.operate.index.ProcessIndex;
-import io.camunda.webapps.schema.descriptors.tasklist.template.TaskTemplate;
+import io.camunda.webapps.schema.descriptors.template.TaskTemplate;
 import io.camunda.webapps.schema.entities.form.FormEntity;
 import java.io.IOException;
 import java.util.List;
