+++ tasklist/importer-870/src/main/java/io/camunda/tasklist/zeebeimport/v870/processors/os/JobZeebeRecordProcessorOpenSearch.java
@@ -18,7 +18,7 @@ import io.camunda.tasklist.util.DateUtil;
 import io.camunda.tasklist.util.OpenSearchUtil;
 import io.camunda.tasklist.zeebeimport.v870.record.Intent;
 import io.camunda.tasklist.zeebeimport.v870.record.value.JobRecordValueImpl;
-import io.camunda.webapps.schema.descriptors.tasklist.template.TaskTemplate;
+import io.camunda.webapps.schema.descriptors.template.TaskTemplate;
 import io.camunda.webapps.schema.entities.usertask.TaskEntity;
 import io.camunda.webapps.schema.entities.usertask.TaskEntity.TaskImplementation;
 import io.camunda.webapps.schema.entities.usertask.TaskState;
