+++ tasklist/importer-common/src/test/java/io/camunda/tasklist/zeebeimport/common/ProcessDefinitionDeletionProcessorTest.java
@@ -16,7 +16,7 @@ import io.camunda.tasklist.store.TaskStore;
 import io.camunda.tasklist.store.VariableStore;
 import io.camunda.webapps.schema.descriptors.index.FormIndex;
 import io.camunda.webapps.schema.descriptors.operate.index.ProcessIndex;
-import io.camunda.webapps.schema.descriptors.tasklist.template.DraftTaskVariableTemplate;
+import io.camunda.webapps.schema.descriptors.template.DraftTaskVariableTemplate;
 import java.util.LinkedHashMap;
 import java.util.List;
 import java.util.Map;
