+++ webapps-schema/src/main/java/io/camunda/webapps/schema/descriptors/template/DraftTaskVariableTemplate.java
@@ -5,7 +5,7 @@
  * Licensed under the Camunda License 1.0. You may not use this file
  * except in compliance with the Camunda License 1.0.
  */
-package io.camunda.webapps.schema.descriptors.tasklist.template;
+package io.camunda.webapps.schema.descriptors.template;
 
 import static io.camunda.webapps.schema.descriptors.ComponentNames.TASK_LIST;
